renamed name.scss to name.module.scss
