+++ packages/terra-clinical-detail-view/src/terra-dev-site/doc/example/DetailViewDivided.module.scss
