+++ packages/terra-clinical-header/src/terra-dev-site/doc/example/ContentHeader.module.scss
