+++ packages/terra-clinical-header/src/terra-dev-site/doc/example/Subheader.jsx
@@ -3,7 +3,7 @@ import Button from 'terra-button';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import Header from 'terra-clinical-header/lib/Header';
 import classNames from 'classnames/bind';
-import styles from './ContentHeader.scss';
+import styles from './ContentHeader.module.scss';
 
 const cx = classNames.bind(styles);
