+++ packages/terra-clinical-label-value-view/src/terra-dev-site/doc/example/LabelValueView.module.scss
